move LevelCompactionPicker to a separate file
