+++ CMakeLists.txt
@@ -485,6 +485,7 @@ set(SOURCES
         db/compaction_job.cc
         db/compaction_picker.cc
         db/compaction_picker_fifo.cc
+        db/compaction_picker_level.cc
         db/compaction_picker_universal.cc
         db/convenience.cc
         db/db_filesnapshot.cc
