+++ TARGETS
@@ -88,6 +88,7 @@ cpp_library(
         "db/compaction_job.cc",
         "db/compaction_picker.cc",
         "db/compaction_picker_fifo.cc",
+        "db/compaction_picker_level.cc",
         "db/compaction_picker_universal.cc",
         "db/convenience.cc",
         "db/db_filesnapshot.cc",
