+++ db/column_family.cc
@@ -21,6 +21,7 @@
 
 #include "db/compaction_picker.h"
 #include "db/compaction_picker_fifo.h"
+#include "db/compaction_picker_level.h"
 #include "db/compaction_picker_universal.h"
 #include "db/db_impl.h"
 #include "db/internal_stats.h"
