+++ src.mk
@@ -12,6 +12,7 @@ LIB_SOURCES =                                                   \
   db/compaction_job.cc                                          \
   db/compaction_picker.cc                                       \
   db/compaction_picker_fifo.cc                                  \
+  db/compaction_picker_level.cc                                 \
   db/compaction_picker_universal.cc                             \
   db/convenience.cc                                             \
   db/db_filesnapshot.cc                                         \
